[evotsc] Add inversion mutational operator
This commit adds inversions to the list of available mutational
operators for Individuals.

To generate an inversion, we pick a start and end point, and replace the
genes between these points in inverse positions, with their orientation
reversed; all the details are explained in the perform_inversion()
method.

Note that if start_pos > end_pos, we still invert between start_pos and
end_pos because it has the same effect (modulo inverting the whole
genome).