[evotsc] Implement 2-environment evaluation
This commit separates the evaluation of individuals in two environments,
A and B.

Each gene now has a type, which represents in which environments it
should be activated: both A and B, only A, or only B.

The difference at evaluation time between environments A and B is given
by their respective values of beta, which is the parameter given to the
activation function:
    activation(beta, x) = tanh(2*(x-1) + beta) + 1

In order to compute the fitness, we compute the expression levels of the
genes in both environments.
Then, we compute the gap in environment X as the average value of each
gene expression level minus its target: 0 if it is supposed to be off in
environment X, and 2 otherwise.

Finally, the fitness is given by:
    f = exp(-selection_coef * (gap_A +gap_B))