[run_evotsc] Barebones runner for evotsc
This commit creates the run_evotsc.py file, which contains a barebones
runner for the evotsc simulation.

It can be called with: `python3 run_evotsc output_dir nb_generations`.

The runner saves the result of plot_expr_AB() and explain() at each save
step (50 by default), and saves the best individual of each save step to
a pickle file.