[evotsc] Add possibility to have several inversions
This commit complexifies the inversion operator a little bit, by adding
the possibility to have several inversions.

Instead of doing an inversion with probability p, we now use p to decide
first if we'll perform mutations.
If we do, we choose the number of mutations according to a geometric
distribution of parameter p.

Taking everything into account, a proportion 1-p of descendants has no
inversion, and the average number of inversions per descendant is 1.