[evotsc] Fix bug in multiple inversions
When several inversions were performed in the same mutation step, the
gene positions were not recomputed, leading to an incorrect handling of
the inversions.

This commit fixes that by always recomputing gene positions when needed
and not storing them in Individuals anymore, as it is not the
time-limiting step in the computation anyway.