[evotsc] Change model to explicit supercoiling
In this commit, we change the heart of the model: the way we compute
expression levels over time.

Now, we compute the local supercoiling for each gene as the sum of the
supercoiling due to three sources:
 - the basal supercoiling
 - the local variation due to TSC
 - the supercoiling variation due to the environment

Then, we pass this value to an activation function (1/(1+e(x)) to obtain
the gene's next expression level.

Minor changes throughout the code also reflect that the expression level
target for activated genes is now 1.0 instead of 2.0.