[evotsc] Fix bug in perform_inversion()
If an inversion started at position 0 and ended after the last gene, the
code didn't recognize that all genes were inverted (therefore meaning
that we don't need to actually do anything).

This resulted in a bug when computing the last intergene distance, which
is fixed by properly taking this case into account.