[evotsc] Make intergene mutations more balanced
This commit tries to fix a bias in intergenic distance mutations.

Up until now, a gene was picked at random, and then a random delta to
add to its intergenic distance.
If the new distance was less than 0, the mutation was not performed at
all.
Since this never happens for delta > 0 but only for delta < 0, this
resulted in a bias towards genome size growth.

This commit tries to remedy this by first picking the delta, and then
picking genes at random until one is found in which the mutation can be
validly performed.

With this modification, the average delta of mutations is really 0, and
only selection determines the final growth rate of the genome.