[evotsc] Reorder population steps
This commit reorders the steps that populations undergo at each
generation.

When stepping, we assume that the population has already been evaluated,
and then pick reproducers, create new individuals with mutations, and
evaluate them.

This ensures that at the end of the generation, the best individual
returned by the step() function is indeed a member of the population of
this generation (and not the one before), and makes it easier to handle
saved populations when analyzing data.