[evotsc] Introduce explicit Euler into run_system
This commit rewrites the run_system() method at the heart of the EvoTSC
model as an explicit Euler ODE solving, using a fixed time step of 1.

It also introduces a new parameter \delta, representing the decay rate
of gene activity, which can be understood as the number of times a gene
has been transcribed recently.