[evotsc/evotsc_run] Add substeps to explicit Euler in run_system()
This commit introduces a more fine-grained control of the explicit Euler
method used to solve the system of gene activity levels.

Instead of an implicit time step of 1, we introduce substeps, that is,
each time step of the lifecycle of the individual is solved by advancing
a number of substeps in the explicit Euler, using a step of 1/substeps.

This greatly improves precision by getting rid of large oscillations
from a substep size of 2; we use a size of 5 for better precision.

This commit also updates evotsc_run.py to take this extra parameter into
account.