[evotsc] Correct error in the model
This commit changes the mathematical model of the TSC underlying the
simulation.

One of the terms used to compute the local supercoiling at a given locus
used the gene activity of other genes, but what it should really use is
the transcription level of the gene (homogeneous to the time derivative
of gene activity).

This commit fixes this error in the model.