[evotsc] Change individual evaluation
In this commit, we replace the evaluation scheme with a new one, not
explicitly time-dependent.

Instead, we consider gene expression levels as a function, of which
we're trying to find a fixed point, using any numerical analysis
algorithm.

After experimental assessments, we chose the scipy.optimize.root method
and the 'linearmixing' method for computing the root of g(x) = f(x)-x,
which gives us the fixed point of f.

As a consequence, when computing the fitness of individuals, we use only
the last time step, as the system is supposed to have converged to a
final static state.