[evotsc] Introduce genes with non-zero length
This commit introduces the concept of gene length (in base pairs).
The idea behind adding a length to genes is that, due to the maximal
interaction distance for the TSC, only a finite number of genes can now
pack together within that distance, as opposed to before where all the
genes could concentrate around the same spot in the genome.

The commit appropriately changes all functions in the evotsc core, as
well as in the evotsc_run module and plotting functions in evotsc.py.