[evotsc] New fitness computation
This commit introduces a new method to compute the fitness of
individuals, based on the final expression level of their genes.

Instead of taking the square of the distance of each gene to its target
individually (and summing over genes), we compute the average distance
of each gene type to its target before taking the square.

This breaks a symmetry that existed when a gene switches from being on
in both environments to being off in both environments: whereas this
previously changed a (1-x)^2 and a x^2 term to x^2 and (1-x)^2,
essentially not changing the fitness, the gene switch now influences the
average expression in both environments: as the average that is most
different to its target is brought closer to it, the fitness gets better.

This should make the fitness landscape less symmetric and therefore
guide evolution better