[evotsc] Change the computation of the distance between genes
With the introduction of non-zero length genes, we started using the
distance between gene promoters as the distance used to compute the
strength of the interaction between genes.

However, a more relevant measure is the distance between the middle of
the transcribed gene and the promoter of the subject gene, as the
average position over time of the polymerase is the middle of the gene
rather than its start.

This commit therefore changes the distance computation in the
compute_inter_matrix() method.