[evotsc{_plot}] Update compute_final_sc_at() for long genes: part I
This commit adds the computation of the SC level due to transcription
at a position that is inside of a gene.

With long genes, supercoiling is maximally negative at the promoter,
maximally positive at the end of the gene, and varies linearly
inbetween.

This commit also fixes an old bug in compute_final_sc_at, which had a
sign error due to incorrect copy/pasting of the compute_inter_matrix
code.
The sign error was being corrected in the plot_genome_and_tsc() method,
but that is not needed anymore.