[evotsc] Switch to double exponential model
This commit tweaks the model to bring it closer to the model described
in the El Houdaigui et al. (2019) paper.

In particular, gene expression e is now computed as the exponential of
promoter activity a, with e = exp(m * a), where a is computed as in the
old model, with a = 1 / (1 + exp(sigma / epsilon)).

In the commit, we also change the default parameter values of epsilon,
sigma_opt, and sigma_basal, and introduce m, to match the model.